requirements.txt
# OCR Bounding Box - Python dependencies
# Install with: pip install -r requirements.txt
# Also requires: sudo apt-get install tesseract-ocr tesseract-ocr-eng

pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python-headless>=4.8.0
numpy>=1.24.0